+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file.
 ### Deprecations
 
 * Core: deprecate our `Cancelled` error type in favour of Swift's built-in `CancellationError` type.
+* Core: due to unuse, we're marking `OptionalIdentifiable` as deprecated.
+* Core: renamed our own `Identifiable` protocol to `TaggedIdentifiable`.
+* CoreData: due to incorrectness, we're deprecating `responseInsert` (this is not `responseInsert`, see below).
 
 ### Improvements
 
@@ -18,6 +21,7 @@ All notable changes to this project will be documented in this file.
 * Core: add `requestVoid` method to network client.
 * Core: default error parser now ignores `.explicitlyCancelled` (for cancelled requests).
 * Core: result "cancellation" now uses Swift's built-in `CancellationError` type if available.
+* Core: `TaggedIdentifiable` protocol now conforms to Swift's own `Identifiable` protocol.
 * XcodeGen: added a template for shared frameworks.
 
 ### Bug Fixes
@@ -25,7 +29,7 @@ All notable changes to this project will be documented in this file.
 * Fastlane: ensure translations export can handle "empty" targets.
 * Scripts: fix Xcode warnings for build steps without input/output files.
 * Bump dependencies (Sentry to 8).
-* CoreData: ensure `requestInsert` imports objects on the context's thread. Note: because of limitations, we're deprecating `responseInsert`.
+* CoreData: ensure `requestInsert` imports objects on the context's thread.
 
 ### Internal